requirements.txt
opencv-python
timm
tqdm
kornia
gdown
transformers
matplotlib
moviepy
pandas
ultralytics
numpy
librosa
pydub
transparent_background
mediapipe
scikit-learn
boto3
ffmpeg
openai-whisper
flask
elevenlabs==0.2.26
anthropic